split client methods into dedicated files; update references
